Write Docstrings on methods. Make some silly unworking test.
