--- admin_panel/project/tests.py
@@ -1,2 +1,0 @@
-import pytest
-
